pyproject.toml
[project]
name = "deep-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain[openai]>=1.2.15",
    "python-dotenv>=1.0.0",
]
